Split Spring and SpringMVC Config
