+++ src/main/java/fr/excilys/databasecomputer/configuration/SpringConfiguration.java
@@ -1,7 +1,5 @@
 package fr.excilys.databasecomputer.configuration;
 
-import java.util.Locale;
-
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRegistration;
@@ -12,21 +10,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
 import org.springframework.core.env.Environment;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.web.WebApplicationInitializer;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 import org.springframework.web.servlet.DispatcherServlet;
-import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.i18n.CookieLocaleResolver;
-import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
-import org.springframework.web.servlet.view.JstlView;
 
 @Configuration
 @EnableWebMvc
@@ -34,7 +23,7 @@ import org.springframework.web.servlet.view.JstlView;
 		"fr.excilys.databasecomputer.mapper","fr.excilys.databasecomputer.service","fr.excilys.databasecomputer.validator",
 		"fr.excilys.databasecomputer.pageable"})
 @PropertySource(value = "classpath:database.properties")
-public class SpringConfiguration implements WebApplicationInitializer, WebMvcConfigurer {
+public class SpringConfiguration implements WebApplicationInitializer {
 
 	@Autowired
 	private Environment env;
@@ -49,58 +38,14 @@ public class SpringConfiguration implements WebApplicationInitializer, WebMvcCon
 		 return dataSource;
 	 }
 
-    @Bean
-    public ViewResolver internalResourceViewResolver() {
-      InternalResourceViewResolver bean = new InternalResourceViewResolver();
-      bean.setViewClass(JstlView.class);
-      bean.setPrefix("/WEB-INF/views/");
-      bean.setSuffix(".jsp");
-
-      return bean;
-    }
-
-    @Override
-    public void addResourceHandlers(final ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/resources/**").addResourceLocations("/resources/");
-    }
-
 	@Override
 	public void onStartup(ServletContext servletContext) throws ServletException {
 		AnnotationConfigWebApplicationContext webContext = new AnnotationConfigWebApplicationContext();
-		webContext.register(SpringConfiguration.class);
+		webContext.register(SpringConfiguration.class,SpringMvcConfiguration.class);
 		webContext.setServletContext(servletContext);
 		ServletRegistration.Dynamic servlet = servletContext.addServlet("dynamicServlet", new DispatcherServlet(webContext));
 		servlet.setLoadOnStartup(1);
 		servlet.addMapping("/");
-	}
-    
-	@Bean
-	public ReloadableResourceBundleMessageSource messageSource() {
-		ReloadableResourceBundleMessageSource ret = new ReloadableResourceBundleMessageSource();
-		ret.setBasename("classpath:configMessages/messages");
-		ret.setDefaultEncoding("utf-8");
-		return ret;
-	}
-	
-	@Bean
-	public CookieLocaleResolver localeResolver(){
-	    CookieLocaleResolver resolver = new CookieLocaleResolver();
-	    resolver.setDefaultLocale(Locale.ENGLISH);
-	    resolver.setCookieName("DatabaseApp");
-	    resolver.setCookieMaxAge(4800);
-	    return resolver;
-	}
-	
-	@Bean
-    public LocaleChangeInterceptor localeChangeInterceptor() {
-        LocaleChangeInterceptor localeChangeInterceptor = new LocaleChangeInterceptor();
-        localeChangeInterceptor.setParamName("local");
-        return localeChangeInterceptor;
-    }
-	
-	@Override
-	public void addInterceptors(InterceptorRegistry registry) {
-		registry.addInterceptor(localeChangeInterceptor());
 	}
 
 }
