+++ src/main/java/fr/excilys/databasecomputer/configuration/SpringMvcConfiguration.java
@@ -1,0 +1,61 @@
+package fr.excilys.databasecomputer.configuration;
+
+import java.util.Locale;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.i18n.CookieLocaleResolver;
+import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+import org.springframework.web.servlet.view.JstlView;
+
+public class SpringMvcConfiguration implements WebMvcConfigurer{
+
+    @Bean
+    public ViewResolver internalResourceViewResolver() {
+      InternalResourceViewResolver bean = new InternalResourceViewResolver();
+      bean.setViewClass(JstlView.class);
+      bean.setPrefix("/WEB-INF/views/");
+      bean.setSuffix(".jsp");
+
+      return bean;
+    }
+
+    @Override
+    public void addResourceHandlers(final ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("/resources/**").addResourceLocations("/resources/");
+    }
+
+    @Bean
+	public ReloadableResourceBundleMessageSource messageSource() {
+		ReloadableResourceBundleMessageSource ret = new ReloadableResourceBundleMessageSource();
+		ret.setBasename("classpath:configMessages/messages");
+		ret.setDefaultEncoding("utf-8");
+		return ret;
+	}
+
+	@Bean
+	public CookieLocaleResolver localeResolver(){
+	    CookieLocaleResolver resolver = new CookieLocaleResolver();
+	    resolver.setDefaultLocale(Locale.ENGLISH);
+	    resolver.setCookieName("DatabaseApp");
+	    resolver.setCookieMaxAge(4800);
+	    return resolver;
+	}
+
+	@Bean
+    public LocaleChangeInterceptor localeChangeInterceptor() {
+        LocaleChangeInterceptor localeChangeInterceptor = new LocaleChangeInterceptor();
+        localeChangeInterceptor.setParamName("local");
+        return localeChangeInterceptor;
+    }
+
+	@Override
+	public void addInterceptors(InterceptorRegistry registry) {
+		registry.addInterceptor(localeChangeInterceptor());
+	}
+}
